model_service/requirements.txt
fastapi
uvicorn[standard]
pandas
scikit-learn
xgboost
psycopg2-binary
joblib
pydantic
